attempt to fix a bug described by a user on the forum regarding PMs
one user on the forum described an error from `hipmost private list':
https://forum.mattermost.org/t/hipchat-importation/5409/9

i think it is probably related to this:
https://stackoverflow.com/q/19549098/2636454

the fix suggested there is to check whether the JSON value is actually
encapsulated in an array as the first element thereof and then to act
accordingly if so. hopefully this fixes the problem.

as to the 2nd error described on the forum, that is currently a puzzle
to me.